networksManager: Adjust to GJS changes
As strings are guaranteed to be UTF-8 in the GNOME platform,
g_file_load_contents() returns raw data instead. Until recently,
GJS used a custom ByteArray type in that case, but with the update
to mozjs60 that was dropped in favor of the native Uint8Array.

As a consequence, we have to explicitly create a string from the
raw data before passing it on to the JSON parser.