build: Include VCS tag in version
This is useful information to include in development snapshots, so use
the output of `git-describe` as version when building from git.

For stable releases there is no change, as they are either build from
a tag (where `git describe` simply returns the tag name) or from a
tarball (where there's a fallback to meson.project_version()).

https://gitlab.gnome.org/GNOME/polari/issues/61