window: Add a visual hint when running a snapshot build
As outlined in the DevOps-with-Flatpak initiative[0], it is important
that development versions are instantly recognizable as such, similar
to the style hint for test instance we added in commit aaec4ee.

However the existing hint is only added when a test instance is actually
running along-side a regular one, so it is not enough to identify a
snapshot build in all cases.

For that, add a different visual hint for snapshots and make sure it
complements the existing hint if both are applied.

[0] https://gitlab.gnome.org/GNOME/Initiatives/wikis/DevOps-with-Flatpak

https://gitlab.gnome.org/GNOME/polari/issues/61