chatEntry: Use GTK's built-in emoji support
Our emoji picker broke in the recent gjs update, as load_contents()
now returns Uint8Array which doesn't map transparently to strings.

Take this as an opportunity to switch to GTK's built-in emoji picker
instead of fixing it - we got our picker slightly ahead of GTK, but
nowadays there's no good reason to keep that duplicated work around.

Current Adwaita is still in need for an Italian job though, so keep
Lapo's original picker styling around for now.

https://gitlab.gnome.org/GNOME/polari/merge_requests/58