CI: Skip enviroments for the GNOME/master branch
We use enviroments only so the review app can link to the generated
flatpak bundle that corresponds to the Merge Requests.
Gitlab environments are not useful if they are created from the
master branch, so there is no need to do so.

Note that flatpak artifacts will continue to be generated for
the master branch. That's a separate task.

https://gitlab.gnome.org/GNOME/Initiatives/wikis/DevOps-with-Flatpak#flatpak-bundle-for-every-mr-and-commit