style: Use consistent spacing in destructured imports
This is a relatively new pattern in our code, which may explain why we
failed so far to adopt a consistent style. Current we omit spaces about
twice as much as we add them, but the latter is consistent with the
style we use for object literals and (non-imports) destructuring, so
go with that.

It's also the style used by Builder's 'gi' snippet, so there's another
precedent.

https://gitlab.gnome.org/GNOME/polari/merge_requests/62