flatpak: Drop workarounds
The first issue was fixed in gjs and is no longer needed.

The second issue has been partially addressed in that the portal
will now fetch the initial state, albeit asynchronously and thus
with a small delay. This is unfortunately going to stay with us,
so we'll need to properly address this in master instead of adding
a partial workaround in the flatpak.

https://gitlab.gnome.org/GNOME/polari/merge_requests/68