networks: Add optional charset field to servers
Not all networks use UTF-8, so any input and output requires conversion
to avoid garbled characters. Telepathy-idle handles that transparently,
provided the correct 'charset' account property is specified.

For predefined networks, we can simply pass on the option if the network
data specifies it. We'll have to see whether it is worth exposing this
also in the UI and thus supporting custom networks as well.

https://gitlab.gnome.org/GNOME/polari/issues/43