app: Do not auto-enable autostart when running in background
Polari is a user application rather than a service, so whether it should
be started at login is better left to the user than tied to the desire
to get its window out of the way.

We also never put the hacks and sandbox holes in place that would have
made it work with flatpak, so this also removes an inconsistency between
distribution channels.

https://gitlab.gnome.org/GNOME/polari/issues/89