mainWindow: Remove close confirmation dialog
Now that the run-in-background setting no longer affects autostart,
the checkbox in the menu is self-explanatory enough and we no longer
need a one-time dialog to explain the feature.

As an added benefit the feature becomes less intrusive, as users
can discover it at their own pace rather than having a choice
forced upon them when using the app for the first time.

https://gitlab.gnome.org/GNOME/polari/issues/89