app: Turn run-in-background feature into an explicit action
As we are reducing the ambiguity with services, it makes more sense
to expose the run-in-background feature as an action the user can
select on a case-by-case basis than having a setting that modifies
the behavior when the window is closed.

This will also remove the need of continuing to expose the quit action
when we drop the app menu, as quit becomes again synonym with closing
the main window.

https://gitlab.gnome.org/GNOME/polari/merge_requests/77