telepathyClient: Unregister on shutdown
So far we simply keep the client registered, as it will eventually
drop off the bus anyway when the application exists. However we want
to allow replacing the running instance on updates, so we should free
up the name as early as possible to allow the new instance to take
over.

https://gitlab.gnome.org/GNOME/polari/merge_requests/80