chatView: Fix warning on shutdown
One reason for a nick status change is that we disconnect the account.
If we also close the corresponding room at the same time (for example
on shutdown), there's a good chance that the room's channel property
is unset after we send the request to ack all pending messages, but
before the operation's callback is called. Avoid accessing an undefined
property in that case by using the callback parameter, which will still
be alive under all circumstances at that point.

https://gitlab.gnome.org/GNOME/polari/merge_requests/79