application: Avoid duplicated ::prepare-shutdown emissions
Since commit 00ab02ca524 added the ability to run in the background,
the ::prepare-shutdown signal is emitted either when the main window
is closed (and the app shouldn't keep running), or when explicitly
quitting the app. But as the quit action also closes all windows, it
is possible for the signal to be emitted twice, resulting in a warning
when the telepathy client releases an additional app reference (it
doesn't hold).

Avoid this by making sure the signal handler which emits the signal
when closing the main window is disconnected from the quit action.

https://gitlab.gnome.org/GNOME/polari/merge_requests/79