room: Add :channel-error property
While we have TpAccount:connection-error to indicate problems with
a network connection as a whole, we don't have any indication for
errors that happen during channel creation (for example a full or
invite-only room).

Add a new :channel-error property that can be used to pass around
such an error condition.

https://gitlab.gnome.org/GNOME/polari/issues/17