chatView: Request log entries on map
If the number of rooms is high enough, querying log entries for all
chatViews at the same time is likely going to result in a bunch of
thread contention (since the DB interfaces are limited).

Make it query on map instead, which will likely be one at a time.

Part-of: <https://gitlab.gnome.org/GNOME/polari/-/merge_requests/264>