lib: Use TrackerResource "add" API when we know replacing is not necessary
On invariable fields, or those we know have no previous values (eg. a
new message as an anonymous node), we can use tracker_resource_add*
API, which avoids deleting previous values. This should speed up
imports.

This is specially important on the "rdf:type" property, as Tracker
implicitly deletes resources when all rdf:types are gone, so we are
essentially deleting and creating elements for channels/users very
frequently.