mainWindow: Avoid warning when no room was selected previously
On startup we first load the previously joined rooms, then select
the last selected one. However the setting for the latter may return
an empty object (when there were no rooms to restore), resulting in
a JS warning about accessing an undefined property.

Add a check to avoid that warning.

https://gitlab.gnome.org/GNOME/polari/merge_requests/100