room: Change highlighting behavior of private chats
We never highlight any messages in private chats, not least
because we assume that the desktop handles notifications for
them. That assumption isn't necessarily true (not even for
GNOME anymore), so this is a good opportunity to reflect what
behavior we actually want for private chats:

 - show notifications for any message (either by the desktop or
   ourselves)

 - highlight any unread message in the sidebar

 - don't highlight any message in the chat log (because a bold
   font weight isn't effective when used for everything)

Only the last point matches what should_highlight() currently does,
however as conceptually *all* private messages are important, it
makes more sense reflect that in the method and special-case the
caller in chatView.

https://gitlab.gnome.org/GNOME/polari/issues/18