joinDialog: Remove obsolete comment
There used to be a vague idea about sorting networks by recency,
frequency or similar, mostly to make it more likely that the
desired network is selected when the dialog is opened.

This was addressed differently by remembering the last used network,
so this is no longer an issue. And the predictable alphabetical order
used now likely makes most sense.

https://gitlab.gnome.org/GNOME/polari/merge_requests/109