style: Always use braces when nesting blocks
We usually avoid braces if the outer block contains a single statement,
except when that would result in ambiguity or worse legibility.
Unfortunately that's a matter of taste that cannot be expressed in an
automated rule. As removing the braces in the cases where we have them
would be worse than adding them where they are avoided, settle
on the latter.

https://gitlab.gnome.org/GNOME/polari/merge_requests/108