style: Only omit braces for single-line statements
We currently omit braces if all blocks consist of a single statement,
but the statement itself may still be spread over multiple lines.
In some places we are stretching it, but in general it's fine. Except
that if we want to automate the rule, eslint forces us to pick between
single-statement without, and single-line with nested handling.

https://gitlab.gnome.org/GNOME/polari/merge_requests/108