mainWindow: Show a clear indication when offline
Polari is almost entirely useless without an internet connection, yet
our current indication that connectivity has been lost is so subtle
that it is easily missed.

To address this, try harder to get the user's attention by popping up
an info bar that outlines the problem and offers a "Network Settings"
button to help them resolve the issue.

https://bugzilla.gnome.org/show_bug.cgi?id=760833