app: Process --version locally
We currently handle the option after registering the application, which
means we will fire up the telepathy client and connect accounts. This
is not the behavior one would expect, which is to simply print the
version and exit.

https://gitlab.gnome.org/GNOME/polari/merge_requests/119


(cherry picked from commit b6a976fccacaa22c17d2e04886e1143ae4251465)