roomList: Always clear connecting indication on status change
When joining a room takes longer than three seconds after establishing
the connection, we show an indication to convey progress to the user.

That indication is hidden again when either the room is joined, or the
connection status changes again. However there is a window that allows
the indicator to get stuck: If the event that hides the indicator happens
*before* the three-second timeout is triggered, then we end up showing
the indicator anyway.

Make sure to cancel the timeout properly in those cases.

https://gitlab.gnome.org/GNOME/polari/merge_requests/120