app: Rearrange code
We want to avoid else blocks that contain a single if statement, as they
can be replaced with a more concise if-else; however here we want to stress
that the code we run when the network is valid or when it becomes valid is
identical. Achieve that while still avoiding a "lonely if" by swapping the
two blocks around.

https://gitlab.gnome.org/GNOME/polari/merge_requests/133