accountsMonitor: Improve reachable check
While we don't block when determining whether a network is reachable
or not, we check servers sequentially until we find one that works,
which is less effective than starting all checks in parallel.

Do that via some Promise black magic that resolves successfully when
the first address can be reached, and rejects when all addresses fail.

https://gitlab.gnome.org/GNOME/polari/merge_requests/132