thumbnailer: Add helper script
We want to handle all types of URL content which means we will
actually have to load the web page and take a snapshot of it.

Embedding a web browser inside Polari isn't great for either
performance or security, so implement the functionality in a
separate script that can run from the main process.

https://gitlab.gnome.org/GNOME/polari/issue/58