urlPreview: Load actual thumbnails
We potentially want to load a lot of previews at a time. Depending on
the web page this can be a quite expensive operation, so we only want
to create one thumbnail at a time.

To ensure that, delegate the loading and creation to a singleton that
runs the previously added helper script on each URL one by one and cache
the result.

https://gitlab.gnome.org/GNOME/polari/issue/58