thumbnailer: Special-case image URLs
Currently we use a generic approach for creating thumbnails which works
well for web pages but can potentially have bad results for smaller
images, as the web view adds a lot of blank space which also appears
in the thumbnail.

To regress this, detect image URLs and load them at the desired
thumbnail size.

https://gitlab.gnome.org/GNOME/polari/issues/58