pasteManager: Catch newly-added gjs exception
gjs 1.58.2 started throwing an exception when encountering a parameter that
is an array of plain structs, as that case is not currently supported.

The exception is overly broad though, and also broke the one case that is
supported: An empty array where the actually struct type isn't relevant.

We cannot fix this in Polari, but at least we can catch the exception to
fail gracefully, so only DND drop target support gets broken instead of
the whole interface.

https://gitlab.gnome.org/GNOME/polari/issues/114