thumbnailer: Add special image handling
Currently we use a generic approach for creating thumbnails which works
well for web pages but can potentially have bad results for smaller
images, as the web view adds a lot of blank space which also appears
in the thumbnail.

Address this by detecting image URLs, fetch the image's position and size,
and use that to clip the created snapshot.

https://gitlab.gnome.org/GNOME/polari/merge_requests/140