thumbnailer: Handle different image sizes
We currently assume that snapshots use the same ratio as the final
previews. While that is correct at the moment, this is about to
change for image URLs, where we want the snapshot to match the
actual image.

Prepare for this by taking the preview size as maximum sizes, and
fit in the source surface as best as possible.

https://gitlab.gnome.org/GNOME/polari/merge_requests/140