connections: Clear passwords when creating account
Telepathy uses a fixed schema for account names which is very prone
to reuse. So when we create a new account, it is quite possible that
there are still left-over credentials from a previous account with
the same name.

Clear those as well to make sure we don't send another network's
passwords to the newly created account.

https://gitlab.gnome.org/GNOME/polari/-/issues/132