entryArea: Fix constant background spinning of spinner
The `GtkSpinner` for showing progress when pasting content to a pastebin
service was always being spun in the background, causing an endless loop
of size requests and style refreshes in GTK which kept Polari using 1–2%
of the CPU.

The `active` property of the spinner was bound to whether the upload
label was visible, and the upload label was visible iff the confirmation
label was *not* visible. However, this was all quite hard to reason
about, and fixing this issue by following that logic would have lead to
calling `this._confirmLabel.show()` once the upload is complete, which
would be quite confusing for someone to read in the future.

Instead, explicitly control the state of the spinner, which is hopefully
a bit more understandable.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://gitlab.gnome.org/GNOME/polari/-/merge_requests/157