thumbnailer: Generate fallback icons on failure
We currently rely on GtkImage to fall back to `image-missing` when the
filename points to a non-existent file. That's not only not the best
fallback, but also means that we try to generate the thumbnail over
and over again.

Avoid this by generating an appropriate fallback icon based on the
guessed content type.

https://gitlab.gnome.org/GNOME/polari/-/issues/149