Disable DropTarget interface temporarily
Drag and drop handling is one of the areas that saw major changes
in GTK4. Those changes are very very welcome, but too major for
a polyfill, and too significant to hide adapting to them in a
general "port to GTK4" commit.

So to allow for a separate port, just stop calling the code for now.

Part-of: <https://gitlab.gnome.org/GNOME/polari/-/merge_requests/232>