roomList: Stop subclassing MenuButton for headers
In GTK4, menu buttons contain either an icon or a label. Prepare
for that by changing the parent class and handling clicks ourselves.

While it seems a bit silly to change the parent to a class that no
longer exists in GTK4, this does help to decrease the diff of the
port.

Part-of: <https://gitlab.gnome.org/GNOME/polari/-/merge_requests/225>