chatView: Stop piggy-backing on LINK state
The GTK4 default theme is a lot more selective on where it applies
a dedicated link style. Prepare for that by defining custom colors
in the application stylesheet instead, and use those consistently
for any elements that represent a nick or link.

Part-of: <https://gitlab.gnome.org/GNOME/polari/-/merge_requests/225>