app: Stop tracking about dialog
The `about` action used to be exposed in the (external) app menu,
so it was reasonable to handle the case where it was activated
with the dialog already showing.

But now that the menu is gone, the only UI to bring up the dialog
is blocked by the dialog itself.

https://gitlab.gnome.org/GNOME/polari/-/merge_requests/174