main: Set up correct typelib/library path
Currently it doesn't matter that the search path we set doesn't match
the install location of the typelib, as imports.package sets up the
correct paths for us later.

However that will stop working when we switch to ES modules for imports,
so prepare for that by setting up the correct paths.

https://gitlab.gnome.org/GNOME/polari/-/merge_requests/176