Split out JS config
We currently set up all build-time variables in config.h, and
pass the ones that are relevant to JS to package.start().

That won't work when we switch to ES modules, as we'll handle
over control to JS by evaluating a (module) file instead of an
embedded script.

Prepare for that by setting up a config.js file with the variables
that are relevant for JS.

https://gitlab.gnome.org/GNOME/polari/-/merge_requests/176