telepathyClient: Fix identifying to NickServ
Commit 042d62b66e introduced an ordering issue in _requestChannel():
When identifying to a bot, the room is only created when observing
the new channel, so looking up the room *before* the request will fail,
even when the request then succeeds. Resetting the channel_error on
the undefined room variable will then throw an exception, which we
silently ignore.

Fix this by moving the channelError handling (including room lookup)
after the channel request completes.

https://gitlab.gnome.org/GNOME/polari/-/merge_requests/195