app: Simplify ::window-removed handling
When running initial setup, we connect a ::window-removed handler to
proceed to the main window when done, then disconnect it and connect
the regular handler when creating the main window.

It is simpler to just use a single handler that covers both cases.

Part-of: <https://gitlab.gnome.org/GNOME/polari/-/merge_requests/225>