app: Use bool instead of time in join-room/message-user
The time parameter in the 'join-room' and 'message-user' actions
is really about whether we should switch to the room or not.

Make this explicit by changing the parameter to a 'present' boolean.
This is not just easier, it also removes the bulk of calls to
`gtk_get_current_event()`, which is gone in GTK4.

https://gitlab.gnome.org/GNOME/polari/-/merge_requests/212