mainWindow: Ensure user-list button is at least as wide as high
For rooms with few users, the button looks too narrow. We used to use
a ::size-allocate hack to make the button at least as wide as icon
buttons, now in GTK4 we can use a cleaner approach and implement a small
layout manager that requests a width that is at least equal to the height.

Part-of: <https://gitlab.gnome.org/GNOME/polari/-/merge_requests/235>