roomList: Restore focus on popover close
If a widget is hidden that contains the focus, GTK will move the focus
(as if TAB was pressed). For popovers in the room list this translates
to switching to the room that is located after the element that popped
up the popover, which is not the behavior we want.

Instead, remember the widget that had the focus when the popover was
opened, and restore it on close.

Part-of: <https://gitlab.gnome.org/GNOME/polari/-/merge_requests/235>