urlPreview: Don't use 'dim-label' on loaded image
For a long time now, the 'dim-label' style has affected the widget's
opacity rather than its (foreground) color. That means it works for
a full-color image just like it does for symbolics, but we really
only want to dim the latter.

Part-of: <https://gitlab.gnome.org/GNOME/polari/-/merge_requests/227>