ircParser: Delegate /NAMES command to user-list popover
The list of users in a room can be fairly large, so popping up
a quick in-app notification with long columns of names for a
couple of seconds isn't great.

The existing user-list popover already provides a cleaner,
searchable UI for the same data, so leverage that instead
of the notification.

Part-of: <https://gitlab.gnome.org/GNOME/polari/-/merge_requests/247>