ircParser: Use feedback popover
The current "command queue" UI is clumsy. A popover seems a better
choice, as it is still more lightweight than a dialog, but also
separated from the chat area.

It is also more flexible, in case we want to improve the layout
with something like title/subtitle later, but for now keep the
existing labels to not introduce string changes late in the cycle.

Part-of: <https://gitlab.gnome.org/GNOME/polari/-/merge_requests/247>