mainWindow: Restructure internal layout
Instead of splitting both titlebar and content area, and synchronizing
the split widths, move the headerbars into the content area and split
that, with a separator in between.

Not only is this simpler because it matches the visual layout, it is
also easier to make adaptive later.

Part-of: <https://gitlab.gnome.org/GNOME/polari/-/merge_requests/250>