messageInfoBar: Replace markup with typography styles
libadwaita defines a set of style classes as part of its API, including
a set of typography styles for labels. Use those instead of pango markup
for better platform consistency.

As a bonus, we can drop the custom setters as we no longer add markup
to the title, so we can simply bind the property.

Part-of: <https://gitlab.gnome.org/GNOME/polari/-/merge_requests/250>