userList: Use automatic accessors
Commit 56efe1afcf77c1e fixed the issue of accidentally overriding
the construct-time value, but now the setters try to access template
children that are only initialized after the GObject constructor.

So solve this differently by replacing the "special" bits from the
setters with a property binding and notify handler. Besides fixing
the warning, this also allows us to rely on gjs' auto accessors.

Part-of: <https://gitlab.gnome.org/GNOME/polari/-/merge_requests/249>