telepathyClient: Show notifications for pending messages
If a private conversation is initiated by another user, the channel
is created in response to a message. That is, the message was received
before the channel existed and is already pending. And as we only
process notifications when receiving a message, we will never show
a notification in that case.

Fix that by explicitly processing pending messages when observing
a new channel.

Part-of: <https://gitlab.gnome.org/GNOME/polari/-/merge_requests/254>