flatpak: Remove access to logs in $HOME
Polari has been distributed as Flatpak for a long time now, so it
is increasingly unlikely that the chat logs are actually shared
between a host- and flatpak instance.

That is, the benefit of not requiring (limited) access to $HOME
now outweighs the ability to easily transition between flatpak
and system package.

Part-of: <https://gitlab.gnome.org/GNOME/polari/-/merge_requests/276>