mainWindow: Add search bar to the sidebar
Add a search bar with a search entry in the sidebar,
this will be used in later commits.

Changes by Carlos Garnacho: Make the search entry to be
in a searchbar, instead of directly in the sidebar box.
The builtin revealer will make showing the entry animatable.
Add ellipsis to placeholder text to follow HIG, make the
placeholder text translatable.

Part-of: <https://gitlab.gnome.org/GNOME/polari/-/merge_requests/311>