style: Make active room labels bold
Now that rows are a bit taller and more generous on spacing, the
visual indication of rooms with unread messages is a bit too soft
on the eye - it merely changes the label color ever so slightly.

Make them a bit more prominent with a bold font variant on rooms
with unread messages.