roomStack: Use AdwToolbarView
By using AdwToolbarView, we get matching overshoot shadows when
scrolling the chat.

Switch to AdwToolbarView, and remove the custom borders applied
to the bottom bar. Bind the bottom bar style to the toolbar style
preferred by the entry area widget, making it raised when the
confirmation is visible, and flat otherwise.

Part-of: <https://gitlab.gnome.org/GNOME/polari/-/merge_requests/285>