join-dialog: Use NavigationView
NavigationView is a new libadwaita widget for page-based navigation,
which is precisely what we are doing in the join dialog.

For now we only replace the outer stack (main and connections),
but keep the inner stack for switching between predefined- and
custom connections. We'll integrate its pages with the new
widgetry in a follow-up change.

Part-of: <https://gitlab.gnome.org/GNOME/polari/-/merge_requests/287>