roomList: Push content on activate, not room change
When the navigation view is collapsed, we currently change
from the sidebar to the contents on room change.

That means it is not possible to get from the sidebar to
the current room's contents without traveling down the
road and back again.

Instead, always switch to the content when a room row is
activated.

https://gitlab.gnome.org/GNOME/polari/-/issues/216

Part-of: <https://gitlab.gnome.org/GNOME/polari/-/merge_requests/292>