lib: Consider TPL_LOG_DIR variable when collecing logs
The flatpak uses the variable to point to the unsandboxed user
dir, to make it possible to pick up logs from the host (when
telepathy-logger is installed and running there).

We shouldn't use logs in that case, so handle the case like
our patched telepathy-logger library did: Use the directory
from the host if it exists, otherwise use the regular user data
dir.

Part-of: <https://gitlab.gnome.org/GNOME/polari/-/merge_requests/264>