chatView: Implement inline display of search logs
Add support for expansion sections in the chatview, so that
we can show inline small pieces of logs from the search view
without pulling all logs in between.

Clicking on the expansion buttons will fetch more logs above
or below it, and expanding it fully will result in the expander
disappearing.

This functionality is hooked to the search view, so clicking
on a result will jump to the chat view and insert related logs
and a expansion section at the appropriate place.

Part-of: <https://gitlab.gnome.org/GNOME/polari/-/merge_requests/311>