joinDialog: Use adaptive dialog
This dialog requires more care, as there's no longer a separate
window that prevents key events from reaching the application
window from processing shortcuts.

Also the dialog is still not fully adaptive, as the connection
selector on the main page is too wide.

It's still a step in the right direction, so no good reason for
delaying it.

Part-of: <https://gitlab.gnome.org/GNOME/polari/-/merge_requests/310>