ci: Only use snapshot builds in merge request pipelines
In order to hook up the new release service, we need a single
job that produces the tarball. This doesn't fit with the
current structure, where we use a separate (non-snapshot)
job for the main branch.

Change this to limit snapshot builds to merge-requests, so
the regular "flatpak" job runs for all branches for which
releases happen.

Part-of: <https://gitlab.gnome.org/GNOME/polari/-/merge_requests/333>