thumbnailer: Operate on Pixbuf instead of cairo surface
GTK4 has reduced the use of Cairo in favor of paintables,
so we would have to create a surface from a paintable,
then convert that to a pixbuf.

That's unnecessarily complex, so to prepare for the port,
create a pixbuf from the returned snapshot immediately
and use that to apply clip and scaling.

Part-of: <https://gitlab.gnome.org/GNOME/polari/-/merge_requests/375>