tools: Add run-eslint.sh script
The script is a small wrapper around the `ci-run-eslint` package
from World/javascript, which is a generalized and improved version
of the .gitlab-ci/run-eslint script we are currently using.

It only requires `npm` to be available, so it simplifies running
eslint locally quite a lot.

Part-of: <https://gitlab.gnome.org/GNOME/polari/-/merge_requests/340>