cleanup: Remove unused eslint-disable directives
The ' camelcase' rule only applies to variables and properties,
not to function names or getter/setters.

The new shared gnome config does not restrict the use of the
imports.format module, so we no longer need an exception to
the 'no-restricted-properties' rule.

Part-of: <https://gitlab.gnome.org/GNOME/polari/-/merge_requests/348>