mainWindow: Disable search mode after showing result inline
When a search result was clicked and is being shown inline, it does
not make much sense to keep the search mode active (and the entry shown,
counters in the room list, filters, ...). Take this as a way to go
back to conversation mode, and untoggle the search mode so the sidebar
updates accordingly.

Part-of: <https://gitlab.gnome.org/GNOME/polari/-/merge_requests/311>