logger: Port to new SparqlConnection singleton
Move away from the singleton handled by libpolari, and into the
new JS singleton. The advantage is that this method is async, so
it can be pulled from async code to avoid blocking.

In order to allow for quickly connecting to channels despite the
store being slow to initialize, also implement some queueing of
inserted data, so it can be handled at once when the connection is
up.

Part-of: <https://gitlab.gnome.org/GNOME/polari/-/merge_requests/365>