utils: Make sure to only create SparqlConnection singleton once
Currently if getSparqlStore() is called again before the previous
call finished, the singleton variable is still unset and so a second
connection will be created.

Avoid this by guarding the operation with a promise, so subsequent
calls wait for the original operation to finish.

Part-of: <https://gitlab.gnome.org/GNOME/polari/-/merge_requests/367>