chatView: Emit signal when logs are ready
The Sparql store can be slow to initialize (e.g. database migrations),
so we should provide feedback in the UI.

To indicate to callers that the chat view's logs are ready, emit a
signal after the initial batch of logs has been loaded and inserted.

Part-of: <https://gitlab.gnome.org/GNOME/polari/-/merge_requests/368>