roomStack: Show spinner while logs are loading
Loading the initial logs is usually fairly quick, but may
take longer if the database has to be migrated after an
update.

To not leave the user with a big empty text view, use a
status page with a spinner as the initial content and only
switch to the chat view when its logs are ready.

Part-of: <https://gitlab.gnome.org/GNOME/polari/-/merge_requests/368>